refactor(core): rename files
